HW5 returned
I fulfilled the question, but there were only two ways I solved the
question from HW4, and we used one for the lecture example last week, so
I used the only other one for HW5, which didn't lend itself well to lots
of methods. I feel that there has to be a way to better unpack
(#NashReference!) the method called populate_hash to make it several
smaller methods, and consequently the whole file stronger and more
flexible.