HW15, disgracefully slapdash
This is literally just the roughest of rough drafts. It all functions,
(except the Pcounter/pcounter_test, which are full of private method
errors and uninitialized constant errors [even THOUGH PUNCT IS
initialized, you stupid thing]) despite the fact that I am sure it wins
no style points, and is kind of clunky.